+++ QtAwesome/QtAwesome.cpp
@@ -215,6 +215,7 @@ const QString QtAwesome::FA_THIN_FONT_FILENAME = "pro/Font Awesome 6 Pro-Thin-10
 const QString QtAwesome::FA_SOLID_FONT_FILENAME = "pro/Font Awesome 6 Pro-Solid-900.otf";
 const QString QtAwesome::FA_SHARP_SOLID_FONT_FILENAME = "pro/Font Awesome 6 Sharp-Solid-900.otf";
 const QString QtAwesome::FA_SHARP_REGULAR_FONT_FILENAME = "pro/Font Awesome 6 Sharp-Regular-400.otf";
+const QString QtAwesome::FA_SHARP_LIGHT_FONT_FILENAME = "pro/Font Awesome 6 Sharp-Light-300.otf";
 #else
 const QString QtAwesome::FA_REGULAR_FONT_FILENAME = "Font Awesome 6 Free-Regular-400.otf";
 const QString QtAwesome::FA_SOLID_FONT_FILENAME = "Font Awesome 6 Free-Solid-900.otf";
@@ -256,6 +257,7 @@ QtAwesome::QtAwesome(QObject* parent)
     _fontDetails.insert(fa::fa_thin, QtAwesomeFontData(FA_THIN_FONT_FILENAME, FA_THIN_FONT_WEIGHT));
     _fontDetails.insert(fa::fa_sharp_solid, QtAwesomeFontData(FA_SHARP_SOLID_FONT_FILENAME, FA_SHARP_SOLID_FONT_WEIGHT));
     _fontDetails.insert(fa::fa_sharp_regular, QtAwesomeFontData(FA_SHARP_REGULAR_FONT_FILENAME, FA_SHARP_REGULAR_FONT_WEIGHT));
+    _fontDetails.insert(fa::fa_sharp_light, QtAwesomeFontData(FA_SHARP_LIGHT_FONT_FILENAME, FA_SHARP_LIGHT_FONT_WEIGHT));
 #endif
 }
 
@@ -319,6 +321,7 @@ bool QtAwesome::initFontAwesome()
     _namedCodepointsByStyle.insert(fa::fa_duotone, _namedCodepointsByStyle.value(fa::fa_solid));
     _namedCodepointsByStyle.insert(fa::fa_sharp_solid, _namedCodepointsByStyle.value(fa::fa_solid));
     _namedCodepointsByStyle.insert(fa::fa_sharp_regular, _namedCodepointsByStyle.value(fa::fa_solid));
+    _namedCodepointsByStyle.insert(fa::fa_sharp_light, _namedCodepointsByStyle.value(fa::fa_solid));
 #else
     addToNamedCodePoints(fa::fa_regular, faRegularFreeIconArray, sizeof(faRegularFreeIconArray)/sizeof(QtAwesomeNamedIcon));
 #endif
@@ -493,6 +496,7 @@ int QtAwesome::stringToStyleEnum(const QString style) const
     else if (style == "fa-duotone") return fa::fa_duotone;
     else if (style == "fa-sharp-solid") return fa::fa_sharp_solid;
     else if (style == "fa-sharp-regular") return fa::fa_sharp_regular;
+    else if (style == "fa-sharp-light") return fa::fa_sharp_light;
 #endif
     return fa::fa_solid;
 }
@@ -509,6 +513,7 @@ const QString QtAwesome::styleEnumToString(int style) const
         case fa::fa_duotone: return "fa-duotone";
         case fa::fa_sharp_solid: return "fa-sharp-solid";
         case fa::fa_sharp_regular: return "fa-sharp-regular";
+        case fa::fa_sharp_light: return "fa-sharp-light";
 #endif
     }
     return "fa_solid";