+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 
-- (2023-02-09) Update to 6.3.0, Aliases and Sharp Regular Pro fonts
+- (2023-03-28) Update to 6.4.0, Sharp Light Pro font
+- (2023-02-09) Update to 6.3.0, Aliases and Sharp Regular Pro font
   - Aliases are now included in the icon list names
   - Support for font style "Font Awesome Sharp Regular" (Pro)
   - Update generator to include all alias names